+++ server/src/provider.js
@@ -50,9 +50,13 @@ Provider.prototype.Clear = function()
 	/** @type {Object.<string, dbInfo>} every key is the lowercase name of db
 	*/
 	this.databases={};
-	/** @type {Array<Group>} An array of groups found*/
+	/** @type {Array<Group>} The array of groups found*/
 	this.groups=[];
 	this.includes=[];
+	/** @type {number[][]} */
+	this.multilineComments=[];
+	/** @type {number} */
+	this.firstLineCommment=-1;
 }
 
 /**
@@ -93,7 +97,7 @@ function Info(name,kind,foundLike,parent,document,startLine,startCol,endLine,end
 	if(comment)
 	{
 		// remove the first newline and replace every character repeated more than 3 times that it is not a space, with 2 of them.
-		this.comment = comment.substr(2).replace(/(\S)\1{2,}/g,"$1$1")
+		this.comment = comment.substr(2).trim().replace(/(\S)\1{2,}/g,"$1$1")
 	}
 }
 
@@ -192,7 +196,7 @@ Provider.prototype.linePrepare = function(line)
 {
 	var justStart = true, precJustStart=true;
 	var precC = " ",c= " ";
-	var string = "";
+	var string = "", stringStart;
 	if(this.currLine.trim().length == 0)
 	{
 		if(line.trim().length == 0)
@@ -204,6 +208,7 @@ Provider.prototype.linePrepare = function(line)
 	{
 		this.currentComment += "\r\n"+this.currLine.trim().substr(4);
 		this.currLine="";
+		if(this.firstLineCommment<0) this.firstLineCommment = this.lineNr;
 		return;
 	}
 	for(var i=0;i<this.currLine.length;i++)
@@ -222,6 +227,7 @@ Provider.prototype.linePrepare = function(line)
 			{
 				if(string=='"e' && precC=='\\')
 					continue; // escaped " inside escaped string
+				this.currLine = this.currLine.substring(0,stringStart+1) + ' '.repeat(i-stringStart-1) + this.currLine.substring(i);
 				string = "";
 			}
 			continue;
@@ -246,6 +252,7 @@ Provider.prototype.linePrepare = function(line)
 					this.currentComment+="\r\n"+this.currLine.substr(i+1)
 					this.comment = true;
 					this.currLine = this.currLine.substr(0,i-1)
+					if(this.firstLineCommment<0) this.firstLineCommment=this.lineNr;
 					return;
 				}
 			}
@@ -254,6 +261,7 @@ Provider.prototype.linePrepare = function(line)
 				// commented line: skip
 				this.currentComment+="\r\n"+this.currLine.substr(i+1)
 				this.currLine="";
+				if(this.firstLineCommment<0) this.firstLineCommment=this.lineNr;
 				return;
 			}
 		}
@@ -264,11 +272,14 @@ Provider.prototype.linePrepare = function(line)
 			this.currentComment+="\r\n"+this.currLine.substr(i+1)
 			this.currLine = this.currLine.substr(0,i-1)
 			this.cont = this.currLine.trim().endsWith(";");
+			if(precJustStart && this.firstLineCommment<0)
+				this.firstLineCommment=this.lineNr;
 			return;
 		}
 		if(c=='"')
 		{
 			string=c;
+			stringStart = i;
 			if(precC=="e")
 			{
 				string+='e';
@@ -278,12 +289,16 @@ Provider.prototype.linePrepare = function(line)
 		if(c=="'")
 		{
 			string=c;
+			stringStart = i;
 			continue;
 		}
 		if(c=="[")
 		{
 			if(/[a-zA-Z0-9_\[]/.test(precC))
+			{
 				string="]";
+				stringStart = i;
+			}
 			continue;
 		}
 	}
@@ -334,6 +349,10 @@ Provider.prototype.parseHarbour = function(words)
 	if(this.currLine.indexOf("#pragma")>=0 && this.currLine.indexOf("BEGINDUMP")>=0)
 	// && /\^s*#pragma\s+BEGINDUMP\s*$/.test(this.currLine)
 	{
+		if(this.currentMethod) {
+			this.currentMethod.endLine = this.lastCodeLine;
+			this.currentMethod = undefined;
+		}
 		this.cMode = true;
 		return;
 	}
@@ -364,6 +383,11 @@ Provider.prototype.parseHarbour = function(words)
 			}
 		}
 	} else
+	if(this.currentClass && (words[0] == "endclass" || (words[0]=="end" && words[1]=="class")))
+	{
+		if(this.currentMethod) this.currentMethod.endLine = this.lastCodeLine;
+		this.currentClass.endLine = this.lineNr;
+	} else
 	if(words[0].length>=4)
 	{
 		if((words[0] == "class") || (words[0]=="create" && words[1]=="class"))
@@ -373,14 +397,6 @@ Provider.prototype.parseHarbour = function(words)
 				this.currentClass = this.addInfo(words[2],'class',"definition")
 			else
 				this.currentClass = this.addInfo(words1,'class',"definition")
-		} else
-		if(words[0] == "endclass")
-		{
-			if(this.currentClass)
-			{
-				if(this.currentMethod) this.currentMethod.endLine = this.lastCodeLine;
-				this.currentClass.endLine = this.lineNr;
-			}
 		} else
 		if(words[0] == "data")
 		{
@@ -493,21 +509,36 @@ Provider.prototype.parse = function(line)
 	this.linePrepare(line);
 	if(this.currLine.trim().length==0 || this.cont) return;
 	/** @type{string[]} */
-	var words = this.currLine.replace(/\s+/g," ").trim().split(" ");
-	if(words.length==0) return;
-	words[0] = words[0].toLowerCase();
-	//console.log((this.cMode?"C":"H")+this.lineNr+">"+this.currLine);
-	if(!this.cMode)
+	if(this.firstLineCommment>=0)
 	{
-		this.findDBReferences();
-		this.parseHarbour(words);
-		if(this.doGroups)
-			this.updateGroups();
-	} else
-	{
-		this.parseC(words);
+		if(this.firstLineCommment<this.lineNr-1) 
+			this.multilineComments.push([this.firstLineCommment,this.lineNr-1])
+		this.firstLineCommment=-1;
 	}
-	this.lastCodeLine = this.lineNr;
+	if(this.cMode) {
+		var words = this.currLine.replace(/\s+/g," ").trim().split(" ");
+		this.parseC(words);
+	} else {
+		var lines = this.currLine.split(";")
+		var pre = ""
+		var code = false;
+		for(var i=0;i<lines.length;i++) {
+			this.currLine = pre+lines[i];
+			//console.debug(this.lineNr+"-"+this.currLine);
+			var words = this.currLine.replace(/\s+/g," ").trim().split(" ");
+			if(words.length==0) continue;
+			code = true;
+			words[0] = words[0].toLowerCase();
+			//console.log((this.cMode?"C":"H")+this.lineNr+">"+this.currLine);
+			this.findDBReferences();
+			this.parseHarbour(words);
+			if(this.doGroups)
+				this.updateGroups(words);
+			pre+=" ".repeat(lines[i].length);
+		}
+	}
+	if(code)
+		this.lastCodeLine = this.lineNr;
 }
 
 /**
@@ -532,6 +563,8 @@ Provider.prototype.parseString = function(txt,docName,cMode)
 Provider.prototype.endParse = function()
 {
 	if(this.currentMethod) this.currentMethod.endLine = this.lastCodeLine;
+	if(this.firstLineCommment>0 && this.firstLineCommment<this.lineNr-1) 
+		this.multilineComments.push([this.firstLineCommment,this.lineNr-1])
 }
 
 /**
@@ -584,15 +617,15 @@ Provider.prototype.findDBReferences = function()
 
 var group_firstWord = /(^|;)\s*([a-z]+(?:\s+(?:[a-z]+))?)/ig;
 var group_keywords = [
-	["if",/#?if(?:n?def)?/,/else(?:if)?/,/end\s*if/],
+	["if",/#?if(?:n?def)?/,/else(?:if)?/,/end\s*(?:if)?/],
 	["for",/for(?:\s+each)?/,"loop","exit","next"],
-	["case",/(switch|do\s+case)/,"case","otherwise","default","exit",/end\s*(switch|case)/],
-	["while",/(?:do\s*)?while/,"loop","exit",/end\s*do/],
-	["try","try","catch",/end\s*do/],
-	["sequence",/begin\s*sequence/,"recover",/end\s*sequence/]
+	["case",/(switch|do\s+case)/,"case","otherwise","default","exit",/end\s*(?:switch|case)?/],
+	["while",/(?:do\s*)?while/,"loop","exit",/end\s*(?:do)?/],
+	["try","try","catch",/end\s*(?:do)?/],
+	["sequence",/begin\s*sequence/,"recover",/end\s*(?:sequence)?/]
 ]; 
 
-Provider.prototype.updateGroups = function()
+Provider.prototype.updateGroups = function(words)
 {
 	var member;
 	var currKeywords;
@@ -603,86 +636,38 @@ Provider.prototype.updateGroups = function()
 		currKeywords = group_keywords.find(v=> v[0]==currGroup.type);
 	}
 
-	while(member=group_firstWord.exec(this.currLine))  {
-		var keyword = member[2].toLowerCase();
-		var pos = this.currLine.indexOf(member[2],member.index);
-		// look for new groups
-		for(var i=0;i<group_keywords.length;i++)
+	var pos = this.currLine.indexOf(words[0]);
+	// look for new groups
+	for(var i=0;i<group_keywords.length;i++)
+	{
+		var m;
+		if((m=words[0].match(group_keywords[i][1])) && m.index==0)
 		{
-			var m;
-			if((m=keyword.match(group_keywords[i][1])) && m.index==0)
-			{
-				currGroup = new Group(group_keywords[i][0]);
-				this.groupStack.push(currGroup);	
-				currGroup.addRange(this.lineNr,pos,pos+m[0].length);
-				currKeywords=group_keywords[i];
-				break;
-			}
+			currGroup = new Group(group_keywords[i][0]);
+			this.groupStack.push(currGroup);	
+			currGroup.addRange(this.lineNr,pos,pos+m[0].length);
+			currKeywords=group_keywords[i];
+			break;
 		}
-		if(currKeywords) for(var i=2;i<currKeywords.length;i++)
+	}
+	if(currKeywords) for(var i=2;i<currKeywords.length;i++)
+	{
+		var m;
+		if((m=words[0].match(currKeywords[i])) && m.index==0)
 		{
-			var m;
-			if((m=keyword.match(currKeywords[i])) && m.index==0)
+			currGroup.addRange(this.lineNr,pos,pos+m[0].length);
+			if(i==currKeywords.length-1)
 			{
-				currGroup.addRange(this.lineNr,pos,pos+m[0].length);
-				if(i==currKeywords.length-1)
+				this.groups.push(this.groupStack.pop());
+				if(this.groupStack.length>0)
 				{
-					this.groups.push(this.groupStack.pop());
-					if(this.groupStack.length>0)
-					{
-						currGroup = this.groupStack[this.groupStack.length-1];
-						currKeywords = group_keywords.find(v=> v[0]==currGroup.type);
-					}
+					currGroup = this.groupStack[this.groupStack.length-1];
+					currKeywords = group_keywords.find(v=> v[0]==currGroup.type);
 				}
-				break;
 			}
+			break;
 		}
 	}
-}
-
-
-Provider.prototype.GetLineAt = function(txt,pos)
-{
-	var lines = [];
-	var posLineStart = pos
-	while(txt[posLineStart]!="\n" && posLineStart>0) posLineStart--;
-	if(txt[posLineStart]=="\n")	posLineStart+=1;
-	var posLineEnd = pos;
-	while(txt[posLineEnd]!="\n" && posLineEnd<txt.length-1) posLineEnd++;
-	// go back until a line without ; found
-	var curStart = posLineStart,curEnd;
-	var line=";"
-	while(line.indexOf(";")>=0 && curStart>0)
-	{
-		curEnd = curStart-1;
-		curStart =  curEnd-1;
-		while(txt[curStart]!="\n" && curStart>0) curStart--;
-		if(txt[curStart]=="\n") curStart+=1;		
-		line = txt.substring(curStart,curEnd)
-		lines.splice(0,0,line)
-	}
-	// go forward until a line without ; found
-	curStart = posLineStart;
-	curEnd = posLineEnd;
-	line = txt.substring(curStart,curEnd)
-	lines.push(line)
-	while(line.indexOf(";")>=0 && curEnd<txt.length-1)
-	{
-		curStart = curEnd+1;
-		curEnd=curStart+1;
-		while(txt[curEnd]!="\n" && curEnd<txt.length-1) curEnd++;
-		line = txt.substring(curStart,curEnd)
-		lines.push(line)	
-	}
-	for(var i=0;i<lines.length;i++)
-	{
-		var line = lines[i];
-		this.linePP(line);
-		if(this.comment || this.cont) contine;
-		this.linePrepare();
-		return this.currLine
-	}
-	return this.currLine
 }
 
 exports.Info = Info;